fix(ListItem): move variables to listItemVariables
