+++ src/components/List/ListItem.tsx
@@ -4,7 +4,7 @@ import * as cx from 'classnames'
 
 import { createShorthandFactory, customPropTypes, pxToRem, UIComponent } from '../../lib'
 import Layout from '../Layout'
-import listVariables from '../../themes/teams/components/List/listVariables'
+import listVariables from '../../themes/teams/components/List/listItemVariables'
 import listItemRules from '../../themes/teams/components/List/listItemRules'
 import { ListItemBehavior } from '../../lib/accessibility'
